fixed sonar violations

+++ src/main/java/com/driw/assingment/shop/driw_shop/dto/ProductDTO.java
@@ -14,5 +14,5 @@ public class ProductDTO {
     private long product_id;
     private String product_name;
     private int itemsPerCarton;
-    private int cartonPrice;
+    private double cartonPrice;
 }
